move the fake class, add some logging
